requirements.txt
Django==2.0
django-crispy-forms==1.7.0
django-pagedown==0.1.3
pytz==2017.3
